requirements-dev.txt
# core
json_strong_typing >= 0.3.2
typing_extensions >= 4.8; python_version<"3.12"

# development tools
build >= 1.0
mypy >= 1.8
flake8 >= 7.0

# PostgreSQL
asyncpg >= 0.29
asyncpg-stubs >= 0.29

# Oracle
oracledb >= 1.4

# MySQL
aiomysql >= 0.2
PyMySQL[rsa]
types-PyMySQL

# Microsoft SQL Server
pyodbc >= 5.0

# Trino
aiotrino >= 0.2
